+++ flytekit/configuration/creds.py
@@ -23,7 +23,7 @@ AUTH_MODE = _config_common.FlyteStringConfigurationEntry("credentials", "auth_mo
 The auth mode defines the behavior used to request and refresh credentials. The currently supported modes include:
 - 'standard' This uses the pkce-enhanced authorization code flow by opening a browser window to initiate credentials
         access.
-- 'basic' This uses cert-based auth in which the end user enters his/her username and password and public key encryption
-        is used to facilitate authentication.
+- 'basic' or 'client_credentials' This uses cert-based auth in which the end user enters a username and password
+        and public key encryption is used to facilitate authentication.
 - None: No auth will be attempted.
 """
